add flag to plugin if entrypoint succeeded and move existing migration to entrypoint
