+++ Viperinius.Plugin.SpotifyImport/ServerEntrypoint.cs
@@ -28,7 +28,19 @@ namespace Viperinius.Plugin.SpotifyImport
         /// <inheritdoc/>
         public Task RunAsync()
         {
-            _logger.LogInformation("Entrypoint");
+            if (Plugin.Instance != null)
+            {
+                try
+                {
+                    RunMigrations();
+
+                    Plugin.Instance.IsInitialised = true;
+                }
+                catch (Exception)
+                {
+                    throw;
+                }
+            }
 
             return Task.CompletedTask;
         }
@@ -46,6 +58,43 @@ namespace Viperinius.Plugin.SpotifyImport
         /// <param name="dispose">Should dispose.</param>
         protected virtual void Dispose(bool dispose)
         {
+        }
+
+        private void RunMigrations()
+        {
+            MigratePlaylistIds();
+        }
+
+        /// <summary>
+        /// Reference: 2023-05-14 [Commit: 0f7374b608f0e2a6287250ffa12731020fb4e40d]
+        /// Convert any existing legacy playlist IDs to full playlist configurations.
+        /// </summary>
+        private void MigratePlaylistIds()
+        {
+            if (!(Plugin.Instance?.Configuration.PlaylistIds.Any() ?? false))
+            {
+                return;
+            }
+
+            var legacyPlaylistCount = Plugin.Instance.Configuration.PlaylistIds.Length;
+
+            var playlists = Plugin.Instance.Configuration.Playlists.ToList();
+            foreach (var playlistId in Plugin.Instance.Configuration.PlaylistIds)
+            {
+                if (!playlists.Where(p => p.Id == playlistId).Any())
+                {
+                    playlists.Add(new Configuration.TargetPlaylistConfiguration()
+                    {
+                        Id = playlistId
+                    });
+                }
+            }
+
+            Plugin.Instance.Configuration.Playlists = playlists.ToArray();
+            Plugin.Instance.Configuration.PlaylistIds = Array.Empty<string>();
+
+            Plugin.Instance.SaveConfiguration();
+            _logger.LogInformation("Migrated {Count} legacy playlist configurations", legacyPlaylistCount);
         }
     }
 }
